edits to test

+++ tests/next_tests/integration_tests/multi_feature_tests/iterator_tests/test_fvm_nabla.py
@@ -383,7 +383,7 @@ def test_nabla_sign(program_processor, lift_mode):
     pnabla_MXX = gtx.as_field([Vertex], np.zeros((setup.nodes_size)))
     pnabla_MYY = gtx.as_field([Vertex], np.zeros((setup.nodes_size)))
 
-    node_sizes = gtx.as_field([], np.asarray((setup.nodes_size)))
+    node_sizes = gtx.as_field([], np.asarray((setup.nodes_size)).astype(np.int32))
 
     e2v = gtx.NeighborTableOffsetProvider(
         AtlasTable(setup.edges2node_connectivity), Edge, Vertex, 2
